# frontend/main/openapi.json
{
  "components": {
    "schemas": {
...
      "AssemblyInput": {
        "additionalProperties": false,
        "description": "The ordered cabinet/filler assembly used to derive the design width.",
        "properties": {
          "cabinets": {
            "items": {
              "$ref": "#/components/schemas/CabinetInput"
            },
            "minItems": 1,
            "title": "Cabinets",
            "type": "array"
          },
          "fillers": {
            "items": {
              "$ref": "#/components/schemas/FillerInput"
            },
            "minItems": 1,
            "title": "Fillers",
            "type": "array"
          }
        },
        "required": [
          "cabinets",
          "fillers"
        ],
        "title": "AssemblyInput",
        "type": "object"
      },
...
      "CabinetInput": {
        "additionalProperties": false,
        "description": "One cabinet in the ordered run the reviewer is checking.",
        "properties": {
          "id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "type": {
            "description": "The reviewer's classification of this cabinet. Slide 11 of the 2026-09-21 deck puts this with the reviewer \u2014 they categorise a cabinet and confirm whether its width may change \u2014 so it is a required input and the server never infers it.",
            "enum": [
              "single_door",
              "double_door",
              "drawer",
              "equipment"
            ],
            "title": "Type",
            "type": "string"
          },
          "width": {
            "description": "Authored dimension with its unit, e.g. 30\" or 762 mm.",
            "maxLength": 100,
            "minLength": 1,
            "title": "Width",
            "type": "string"
          }
        },
        "required": [
          "id",
          "width",
          "type"
        ],
        "title": "CabinetInput",
        "type": "object"
      },
      "CabinetProposalOut": {
        "description": "One cabinet before and after the distribution calculation.",
        "properties": {
          "adjustable": {
            "title": "Adjustable",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "original": {
            "$ref": "#/components/schemas/app__schemas__distribution__QuantityOut"
          },
          "proposed": {
            "$ref": "#/components/schemas/app__schemas__distribution__QuantityOut"
          },
          "type": {
            "enum": [
              "single_door",
              "double_door",
              "drawer",
              "equipment"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "original",
          "proposed",
          "adjustable"
        ],
        "title": "CabinetProposalOut",
        "type": "object"
      },
...
      "FillerDistributionRequest": {
        "additionalProperties": false,
        "description": "Inputs for the deterministic filler-first distribution proposal.\n\n``field_width`` is nullable on purpose: a missing on-site field dimension is a business\nabstention (NOT_FOUND), not a guessed zero or a selected default.",
        "properties": {
          "assembly": {
            "$ref": "#/components/schemas/AssemblyInput"
          },
          "double_door_cab_width_max": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Double Door Cab Width Max",
            "type": "string"
          },
          "double_door_cab_width_min": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Double Door Cab Width Min",
            "type": "string"
          },
          "drawer_cab_width_max": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Drawer Cab Width Max",
            "type": "string"
          },
          "drawer_cab_width_min": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Drawer Cab Width Min",
            "type": "string"
          },
          "field_width": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reviewer-entered site dimension with its unit, e.g. 96 1/2\".",
            "title": "Field Width"
          },
          "filler_max": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Filler Max",
            "type": "string"
          },
          "filler_min": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Filler Min",
            "type": "string"
          },
          "single_door_cab_width_max": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Single Door Cab Width Max",
            "type": "string"
          },
          "single_door_cab_width_min": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Single Door Cab Width Min",
            "type": "string"
          }
        },
        "required": [
          "assembly",
          "filler_min",
          "filler_max",
          "single_door_cab_width_min",
          "single_door_cab_width_max",
          "double_door_cab_width_min",
          "double_door_cab_width_max",
          "drawer_cab_width_min",
          "drawer_cab_width_max"
        ],
        "title": "FillerDistributionRequest",
        "type": "object"
      },
      "FillerDistributionResponse": {
        "description": "A deterministic proposal, or an honest abstention.",
        "properties": {
          "cabinets": {
            "items": {
              "$ref": "#/components/schemas/CabinetProposalOut"
            },
            "title": "Cabinets",
            "type": "array"
          },
          "cabinets_retained": {
            "title": "Cabinets Retained",
            "type": "boolean"
          },
          "calculation": {
            "title": "Calculation",
            "type": "string"
          },
          "condition": {
            "title": "Condition",
            "type": "string"
          },
          "design_width": {
            "$ref": "#/components/schemas/app__schemas__distribution__QuantityOut"
          },
          "field_dimension": {
            "$ref": "#/components/schemas/OperandTraceOut"
          },
          "fillers": {
            "items": {
              "$ref": "#/components/schemas/FillerProposalOut"
            },
            "title": "Fillers",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "operands": {
            "items": {
              "$ref": "#/components/schemas/OperandTraceOut"
            },
            "title": "Operands",
            "type": "array"
          },
          "outcome": {
            "enum": [
              "PASS",
              "REVIEW_REQUIRED",
              "NOT_FOUND"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "reviewer_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Action"
          },
          "site_difference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/app__schemas__distribution__QuantityOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "outcome",
          "condition",
          "message",
          "summary",
          "design_width",
          "site_difference",
          "field_dimension",
          "fillers",
          "cabinets",
          "cabinets_retained",
          "operands",
          "calculation"
        ],
        "title": "FillerDistributionResponse",
        "type": "object"
      },
      "FillerInput": {
        "additionalProperties": false,
        "description": "One filler, ordered left then right in the request.",
        "properties": {
          "id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "width": {
            "description": "Authored dimension with its unit, e.g. 2\" or 51 mm.",
            "maxLength": 100,
            "minLength": 1,
            "title": "Width",
            "type": "string"
          }
        },
        "required": [
          "id",
          "width"
        ],
        "title": "FillerInput",
        "type": "object"
      },
      "FillerProposalOut": {
        "description": "One filler before and after the distribution calculation.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "original": {
            "$ref": "#/components/schemas/app__schemas__distribution__QuantityOut"
          },
          "proposed": {
            "$ref": "#/components/schemas/app__schemas__distribution__QuantityOut"
          }
        },
        "required": [
          "id",
          "original",
          "proposed"
        ],
        "title": "FillerProposalOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperandTraceOut": {
        "description": "The source record the response gives for an input value.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "enum": [
              "ARCH",
              "SHOP",
              "USER_INPUT",
              "LITERAL"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "const": "HUMAN_CONFIRMED",
            "title": "Status",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/app__schemas__distribution__QuantityOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "source",
          "status",
          "value"
        ],
        "title": "OperandTraceOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "app__schemas__distribution__QuantityOut": {
        "description": "One exact dimension, rendered without a JSON float.",
        "properties": {
          "as_typed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Typed"
          },
          "denominator": {
            "title": "Denominator",
            "type": "string"
          },
          "display": {
            "title": "Display",
            "type": "string"
          },
          "numerator": {
            "title": "Numerator",
            "type": "string"
          },
          "unit": {
            "enum": [
              "in",
              "mm"
            ],
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "numerator",
          "denominator",
          "unit",
          "display"
        ],
        "title": "QuantityOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AI-assisted shop drawing review. The AI reads; deterministic Python decides; a reviewer signs off.",
    "title": "Graniti Vicentia V1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/filler-distribution": {
      "post": {
        "description": "Return the two-step proposal, abstaining when reviewer input is missing.",
        "operationId": "calculate_filler_distribution_api_v1_projects__project_id__filler_distribution_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FillerDistributionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FillerDistributionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate the site-corrected layout for one cabinet run",
        "tags": [
          "distribution"
        ]
      }
    },
...
  }
}